third_party/WebKit/LayoutTests/shadow-dom/crashes/css-focus-recalc.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<html>
<body>
<style>
#host:not(:focus)::after {
    content: 'text';
}
</style>

<div id="host" tabindex=0></div>
</body>
<script>
test(() => {
    host.attachShadow({mode: 'closed'});
    host.focus();
    assert_true(true, "This test should not crash.");
}, "the style is recalculated by host.focus()");
</script>
</html>
